+++ shared/node.go
@@ -4,8 +4,8 @@ import (
 	"fmt"
 	"log"
 	"pocket/consensus"
+	"pocket/p2p/pre2p"
 	"pocket/persistence"
-	"pocket/pre2p"
 	"pocket/shared/config"
 	"pocket/shared/crypto"
 	"pocket/shared/types"
